Add optional parameter support

+++ amazons3/tests/test.bal
@@ -25,41 +25,49 @@ string testAccessKeyId = config:getAsString("ACCESS_KEY_ID");
 string testSecretAccessKey = config:getAsString("SECRET_ACCESS_KEY");
 string testRegion = config:getAsString("REGION");
 string testBucketName = config:getAsString("BUCKET_NAME");
-string amazonHost = config:getAsString("AMAZON_HOST");
 
 AmazonS3Configuration amazonS3Config = {
     accessKeyId: testAccessKeyId,
     secretAccessKey: testSecretAccessKey,
-    region: testRegion,
-    amazonHost: amazonHost
+    region: testRegion
 };
 
-Client amazonS3Client = new(amazonS3Config);
-
 @test:Config
-function testGetBucketList() {
-    log:printInfo("amazonS3ClientForGetBucketList -> getBucketList()");
-    var response = amazonS3Client -> getBucketList();
-    if (response is error) {
-        test:assertFail(msg = <string>response.detail().message);
+function testListBuckets() {
+    log:printInfo("amazonS3ClientForGetBucketList -> listBuckets()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        var response = amazonS3Client -> listBuckets();
+        if (response is error) {
+            test:assertFail(msg = <string>response.detail().message);
+        } else {
+            string bucketName = response[0].name;
+            test:assertTrue(bucketName.length() > 0, msg = "Failed to call getBucketList()");
+        }
     } else {
-        string bucketName = response[0].name;
-        test:assertTrue(bucketName.length() > 0, msg = "Failed to call getBucketList()");
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
     }
+    
 }
 
 @test:Config {
-    dependsOn: ["testGetBucketList"]
+    dependsOn: ["testListBuckets"]
 }
 function testCreateBucket() {
     log:printInfo("amazonS3Client -> createBucket()");
-    var response = amazonS3Client -> createBucket(testBucketName);
-    if (response is Status) {
-        boolean bucketStatus = response.success;
-        test:assertTrue(bucketStatus, msg = "Failed createBucket()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        CannedACL cannedACL = ACL_PRIVATE;
+        var response = amazonS3Client -> createBucket(testBucketName, cannedACL = cannedACL);
+        if (response is Status) {
+            boolean bucketStatus = response.success;
+            test:assertTrue(bucketStatus, msg = "Failed createBucket()");
+        } else {
+            test:assertFail(msg = <string>response.detail().message);
+        }
     } else {
-        test:assertFail(msg = <string>response.detail().message);
-    }
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
+    }  
 }
 
 @test:Config {
@@ -67,13 +75,18 @@ function testCreateBucket() {
 }
 function testCreateObject() {
     log:printInfo("amazonS3Client -> createObject()");
-    var response = amazonS3Client -> createObject(testBucketName, "test.txt","Sample content");
-    if (response is Status) {
-        boolean objectStatus = response.success;
-        test:assertTrue(objectStatus, msg = "Failed createObject()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        var response = amazonS3Client -> createObject(testBucketName, "test.txt","Sample content");
+        if (response is Status) {
+            boolean objectStatus = response.success;
+            test:assertTrue(objectStatus, msg = "Failed createObject()");
+        } else {
+            test:assertFail(msg = <string>response.detail().message);
+        }
     } else {
-        test:assertFail(msg = <string>response.detail().message);
-    }
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
+    } 
 }
 
 @test:Config {
@@ -81,40 +94,55 @@ function testCreateObject() {
 }
 function testGetObject() {
     log:printInfo("amazonS3Client -> getObject()");
-    var response = amazonS3Client->getObject(testBucketName, "test.txt");
-    if (response is S3Object) {
-        string content = response.content;
-        test:assertTrue(content.length() > 0, msg = "Failed to call getObject()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        var response = amazonS3Client->getObject(testBucketName, "test.txt");
+        if (response is S3Object) {
+            string content = response.content;
+            test:assertTrue(content.length() > 0, msg = "Failed to call getObject()");
+        } else {
+            test:assertFail(msg = <string>response.detail().message);
+        }
     } else {
-        test:assertFail(msg = <string>response.detail().message);
-    }
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
+    } 
 }
 
 @test:Config {
     dependsOn: ["testGetObject"]
 }
-function testGetAllObjects() {
-    log:printInfo("amazonS3Client -> getAllObjects()");
-    var response = amazonS3Client -> getAllObjects(testBucketName);
-    if (response is error) {
-        test:assertFail(msg = <string>response.detail().message);
+function testListObjects() {
+    log:printInfo("amazonS3Client -> listObjects()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        var response = amazonS3Client -> listObjects(testBucketName, fetchOwner = true);
+        if (response is error) {
+            test:assertFail(msg = <string>response.detail().message);
+        } else {
+            test:assertTrue(response.length() > 0, msg = "Failed to call listObjects()");
+        }
     } else {
-        test:assertTrue(response.length() > 0, msg = "Failed to call getAllObjects()");
-    }
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
+    } 
 }
 
 @test:Config {
-    dependsOn: ["testGetAllObjects"]
+    dependsOn: ["testListObjects"]
 }
 function testDeleteObject() {
     log:printInfo("amazonS3Client -> deleteObject()");
-    var response = amazonS3Client -> deleteObject(testBucketName, "test.txt");
-    if (response is Status) {
-        boolean objectStatus = response.success;
-        test:assertTrue(objectStatus, msg = "Failed deleteObject()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        var response = amazonS3Client -> deleteObject(testBucketName, "test.txt");
+        if (response is Status) {
+            boolean objectStatus = response.success;
+            test:assertTrue(objectStatus, msg = "Failed deleteObject()");
+        } else {
+            test:assertFail(msg = <string>response.detail().message);
+        }
     } else {
-        test:assertFail(msg = <string>response.detail().message);
-    }
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
+    } 
 }
 
 @test:Config {
@@ -122,11 +150,16 @@ function testDeleteObject() {
 }
 function testDeleteBucket() {
     log:printInfo("amazonS3Client -> deleteBucket()");
-    var response = amazonS3Client -> deleteBucket(testBucketName);
-    if (response is Status) {
-        boolean bucketStatus = response.success;
-        test:assertTrue(bucketStatus, msg = "Failed deleteBucket()");
+    AmazonS3Client|error amazonS3Client = new(amazonS3Config);
+    if (amazonS3Client is AmazonS3Client) {
+        var response = amazonS3Client -> deleteBucket(testBucketName);
+        if (response is Status) {
+            boolean bucketStatus = response.success;
+            test:assertTrue(bucketStatus, msg = "Failed deleteBucket()");
+        } else {
+            test:assertFail(msg = <string>response.detail().message);
+        }
     } else {
-        test:assertFail(msg = <string>response.detail().message);
-    }
+        test:assertFail(msg = <string>amazonS3Client.detail().message);
+    } 
 }
